backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery==5.4.0
redis==5.2.0
icalendar==6.1.0
pyotp==2.9.0
qrcode==8.2
python-multipart==0.0.18
httpx==0.28.0

# Feiertage (optional, Fallback ist eingebaut)
# workalendar==17.0.0

# Benachrichtigungen (Phase 2)
# sendgrid==6.11.0
# python-telegram-bot==21.9
# matrix-nio==0.25.2

# PDF (Phase 2)
# reportlab==4.2.5
